done: 104; added: stdint and inttypes for representation; comments: hoefully C90 allows stdint and type reps

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -1,5 +1,7 @@
 #include "main.h"
 #include <stdio.h>
+#include <stdint.h>
+#include <inttypes.h>
 
 /**
  *  fibonacci_98 - finds the first 98 fibo nums from 1, 2.. and prints them.
@@ -8,11 +10,11 @@
  */
 void fibonacci_98(void)
 {
-	unsigned long long int fib_num = 0;
+	int64_t fib_num = 0;
 
-	unsigned long long int prevnum_a = 1;
+	int64_t prevnum_a = 1;
 
-	unsigned long long int prevnum_b = 0;
+	int64_t prevnum_b = 0;
 
 	int i = 0;
 
@@ -25,9 +27,9 @@ void fibonacci_98(void)
 		prevnum_a = fib_num;
 
 		if (i == 97)
-			printf("%lld\n", fib_num);
+			printf("%" PRId64 "\n", fib_num);
 		else
-			printf("%lld, ", fib_num);
+			printf("%" PRId64 ", ", fib_num);
 		i++;
 	}
 }
